ENH: Extensions can be specified in the JSON file
Extensions that are required to create the custom version of Slicer can be specified in the JSON file. If the extensions are not installed, they will be automatically installed. If they are installed, the script will verify if an update is available and warn the user about it (updates have to be installed manually). If an extension is installed but not required for the custom Slicer, it will simply be ignored and not copied in the new package.