BUG: Wrong launcher modified on Linux and Windows
CustomSlicerGenerator.py was modifying SlicerLauncherSettings.ini instead of modifying Slicer[custom]LauncherSettings.ini .
Therefore the packaged launcher that was used (Slicer[custom]LauncherSettings.ini) was left unmodified. This was leading to Slicer[custom] changing the extension paths for the original Slicer on which the custom version was based. This was not a problem if the user was only using Slicer[custom], but it was a problem if the user was using both Slicer and Slicer[custom] with the same revision.