ENH: Deactivate extensions by default with CustomSlicer
To avoid problems, by default, extensions are deactiavated in the newly created custom Slicer. It will also deactivate the extensions each time the custom Slicer is moved on your system. Additionally, the path in which the extensions are installed is also modified to point to the custom Slicer directory to avoid pointing to a directory in which the user may have extensions installed from the same revision of Slicer.