ENH: Custom Slicer name contains version number
This allows one to generate multiple Custom Slicer with the same name, based on the same revision of Slicer, without having to worry about conflicts between configuration files. Before that, the configuration files would have been the same if multiple custom Slicer were based on the same revision of Slicer, which would have created conflicts.
"." have to be removed from the version string when added to the custom Slicer name because CTKLauncher would interpret the "." as the beginning of the extension of the executable file. "-" are used for the name of most files instead, but for the Customizer.py script, "_" had to be used because "-" and "." in the python script name and python class name would create errors.