ENH: detection of error if downloadAndInstallExtension() fails.
Since Slicer r24666 [1], the extension manager function returns a boolean if it fails downloading the metadata associated with the requested extension. This allows to detect if the requested extension is not going to be downloaded and to stop the creation of the customized version of Slicer at that point. The user can then try again (most of the time it is just because the server is too slow to answer) or not by clicking again on 'generate'.
Previously, this error could not be detected. This was resulting in not resetting the 'generate' button. The user had to press 'cancel' even though the creation of the customized Slicer had already failed. This is not necessary anymore.

[1] https://github.com/Slicer/Slicer/commit/d42d6c9ec7e7bb651170ada41e39d717bba57662